move format-secret to vm module
